+++ core/src/main/java/io/questdb/cutlass/http/HttpServerConfigurationWrapper.java
@@ -51,7 +51,7 @@ public class HttpServerConfigurationWrapper implements HttpFullFatServerConfigur
     }
 
     @Override
-    public int getAcceptLoopTimeout() {
+    public long getAcceptLoopTimeout() {
         return getDelegate().getAcceptLoopTimeout();
     }
